Purge post collection endpoints more liberally
When a post is updated, we don't know how its new values will affect the
collections it's represented in. As such, we need to purge all possible
collections, instead of the collections it was previously present in.

Behavior only affects updates to published posts, because the assumption
is that unpublished posts won't ever be cached.