Change the Site Health label to Pantheon (#339)
This commit changes the label for the Site Health test to Pantheon, so it's clear where the test is coming from. It's often difficult to distinguish what tests are coming from WordPress core and what tests are coming from plugins. We can disambiguate that by being clear about tests that we are specifically adding to the Site Health page.